refactor: prep structure for themes
