+++ src/components/themes/default/AlertAreaBar.tsx
@@ -1,6 +1,6 @@
 import React, { useRef, useEffect, useMemo, useImperativeHandle, forwardRef } from "react";
-import { isZoneBased, getCounties, getStates, getCountiesWithStates } from "../../utils/nwsAlertUtils";
-import { colorMap, TAILWIND_TO_HEX } from "../../config/alertConfig";
+import { isZoneBased, getCounties, getStates, getCountiesWithStates } from "../../../utils/nwsAlertUtils";
+import { colorMap, TAILWIND_TO_HEX } from "../../../config/alertConfig";
 
 // Utility to lighten a hex color by a given percent (0-100)
 function lightenHexColor(hex: string, percent = 20): string {
