+++ src/components/themes/default/AlertTypeBar.tsx
